Make the army continuous, and price every gate as a share of it
The report was "now every round the crowd shrinks to minimum... once you achieved
a certain level or crowd size you never lose it", alongside "at the same time I
want it to become more difficult", and "I want shield and spell to work against
all, including gates can be eliminated by the spell... a magazine of spells
instead of one at a time".

Carrying the army between rounds is one line. Shipping that line alone would have
destroyed the game, and the simulation said so before any of this was written.

WALKING THE REAL GENERATOR WITH THE OLD ABSOLUTE GATES, an army carried over is
pinned at the 100,000 soft cap by ROUND FIVE, and every round after it is a flat
1.00x -- every gate in the game doing nothing measurable. The cap is not even the
whole problem: because an add gate paid an absolute headcount, a round entered
with 2,461 men is worth 1.9x against the 45x the same round pays at five men.
Long before the ceiling, the gate stops being an event.

So a gate is a SHARE of the army that meets it. A recruit is x1.026 per weight, a
rally x1.30, an ambush x0.875 at the top of a round and x0.409 at the bottom of a
deep one -- and all of them mean exactly as much at fifty men as at fifty
billion. The player still sees headcounts, because GateMath.Headcount resolves
the share against the army actually approaching: the sign still reads "+340" and
"-1.3K" exactly as before. The soft cap is deleted; nothing needs capping when
nothing is absolute.

THE x2 ARCH DID NOT SURVIVE, AND THAT WAS MEASURED TOO. Keeping literal doubling
and simulating sixty rounds, every lane-choice quality from 0.70 to 1.00 lands
within one order of magnitude of the same colossal number: rally gates dominate
so completely that nothing else the player does is detectable. Skill expression
and the x2 arch are the same trade, and the arch lost. It is x1.30 -- still worth
nearly seven recruit gates, still the gate you steer for.

FORCE IS A DOUBLE, not a long. At the measured growth of a competent player a
long overflows around round 52, which is INSIDE the sixteen authored acts, so
this is not a theoretical ceiling. A double holds the same campaign with two
hundred rounds to spare and needs no saturation reasoning at all.

StandingArmy holds both halves of the promise, and neither number does it alone:
Banked carries unbroken from the last round, BestEver is a high-water mark that
only rises, and the floor is 55% of it. A disastrous round costs up to 45% of a
career; no sequence of them can put the player back at the beginning. A floor of
100% was considered and rejected -- that is not an RPG level, it is an
invulnerability, and the difficulty asked for alongside would have nowhere to
land.

Measured across all 62 rounds from the shipped code, break-even sits at lane
quality 0.72. Below it the army does not grow -- which is the difficulty arriving
where it was aimed, since the ambush weight and the depth ramp climb while the
recruit share does not -- and the floor is what stops that being a spiral. At
0.85 the army reaches 574M by round 30 and 20.1T by round 62, rank 42.

THE BOSS IS PRICED AGAINST THE ARMY THAT ACTUALLY WALKS IN, through the same
CrowdFactor its damage is multiplied by, so the two cancel exactly: the fight
lasts about as long whether the player arrives with four hundred men or four
hundred trillion. BossSim.ExpectedForceAtAct is deleted, because with a
continuous army no modelled ladder can be fair to a player at 0.80 and one at
0.90 -- they end eight orders of magnitude apart, and a boss priced between them
is a formality for one and a wall for the other. Gear and talents stay outside
the model, so they still show as a shorter fight. Bases x0.60 and pressures eased
to 0.036-0.051; the campaign measures 17s rising to 51s across sixteen acts with
the affixes folded in.

TWO BUGS BEHIND "THE SPELL DOESN'T DESTROY ENEMY PACKS AND SHIELD DOESN'T BLOCK
AGAINST THEM", and the first one is arithmetic rather than logic. The sweep
started at the crowd's CENTRE and ran fifteen metres, while the crowd's own front
line already stands up to seven metres ahead of that centre -- so it cleared
about eight metres of road, under one second of travel at the run speed, which is
less time than it takes to see an ambush and flick at it. It was not a
short-ranged spell; it was a spell that could only hit what was already
unavoidable. The second is that it iterated enemy PACKS only, while most of the
red on the road is a subtract GATE, drawn as a crowd of men since v0.20 and
walking straight through a spell aimed at it. Both now act on ambushes of either
kind, from the army's front, at 34 m. A raised shield nullifies either, and says
so: blocking used to cost nothing and show nothing, which -- against the old
absolute pack cost of a few dozen men to an army of hundreds -- was a difference
too small to notice even when it was NOT blocked.

BOTH ABILITIES ARE MAGAZINES. Core/Run/Magazine refills whenever it is short
rather than only when empty, and that rule is the whole reason a second charge is
worth a talent point: with three charges a player can spend two on a dangerous
chunk and still have the third when the next one arrives. Widening the magazine
does not hand out the new charges, or equipping a charge talent mid-run would be
a free cast. New SpellCharges and ShieldCharges stats, both counts of EXTRA
charges so a player who has bought none gets exactly the ability that shipped
before, bought by Full Quiver / Arsenal and Doubleguard / Triguard. The keystone
payoff is folded into the EXISTING keystones rather than added as a fourth --
three mutually exclusive keystones per branch is a design rule with a test on it,
and quietly making it four to fit a new stat in would change the shape of every
build in the game to avoid an edit.

Seven pinned tests are DELIBERATELY REWRITTEN and each says why in its own body.
Every one of them asserted something about absolute arithmetic or integer
rounding -- a +5 adds five men, a clean 40% of 1000 removes exactly 400 whole
units, a pack always bites for at least one. Those were true statements about a
game that re-mustered five men a round, and they are meaningless now. What each
was really protecting is re-expressed rather than dropped.

Also: par collapses below 1.0 from about round twenty, which is the same fact
seen from the other end, so revives are sized against the army that walked in
rather than against par -- a revive sized off par would have handed a deep-run
player fewer men than they started the round with.

+++ Assets/Scripts/Core/Boss/BossMelee.cs
@@ -115,7 +115,7 @@ namespace BattleRunner.Core.Boss
         /// appeared. Always at least one while the army exists: a boss fight with nobody
         /// visibly fighting is the bug this class removes.
         /// </summary>
-        public static int Fighters(long force)
+        public static int Fighters(double force)
         {
             if (force <= 0) return 0;
             double want = 4.0 + Math.Sqrt(force) * 1.15;
